requirements.txt
numpy
pytorch==1.2.0
matplotlib==3.1.1
tqdm
torchtext==0.4.0
torchvision